feat(identity): extract compatible auth domain
