+++ .github/CODEOWNERS
@@ -18,6 +18,7 @@
 /booking/ @jubenitogarcia
 /crm/ @jubenitogarcia
 /finance/ @jubenitogarcia
+/identity/ @jubenitogarcia
 /integration/ @jubenitogarcia
 /inventory/ @jubenitogarcia
 /messaging/ @jubenitogarcia
