+++ identity/package.json
@@ -1,0 +1,8 @@
+{
+  "name": "@skincos/identity",
+  "private": true,
+  "type": "module",
+  "scripts": {
+    "test": "node --test test/*.test.mjs"
+  }
+}
