+++ inventory/package.json
@@ -7,7 +7,8 @@
     "scripts": {
         "dev": "wrangler dev --config wrangler.toml",
         "deploy": "wrangler deploy --config wrangler.toml --keep-vars",
-        "d1:migrate": "wrangler d1 migrations apply skincos-db --config wrangler.toml",
+    "d1:migrate": "wrangler d1 migrations apply skincos-db --config wrangler.toml",
+    "test": "node --test tests/*.test.mjs",
         "lint": "echo \"(todo)\""
     },
     "devDependencies": {
