+++ shared/crm-auth/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "@skincos/crm-auth-adapter",
+  "private": true,
+  "type": "module"
+}
