+++ shared/identity-contract/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "@skincos/identity-contract",
+  "private": true,
+  "type": "module"
+}
